Add button for inviting staff users
This takes the user to the form which they can use to invite a new
member of staff to the application.
