+++ app/controllers/staff/invitations_controller.rb
@@ -38,7 +38,11 @@ class Staff::InvitationsController < Devise::InvitationsController
     end
   end
 
+  def after_invite_path_for(inviter, invitee)
+    invitee.is_a?(Staff) ? support_interface_staff_index_path : super
+  end
+
   def after_accept_path_for(resource)
-    resource.is_a?(Staff) ? support_interface_root_path : super
+    resource.is_a?(Staff) ? support_interface_staff_index_path : super
   end
 end
