+++ app/views/support_interface/staff/index.html.erb
@@ -1,3 +1,7 @@
 <% content_for :page_title, 'Staff' %>
 
 <h1 class="govuk-heading-l">Staff</h1>
+
+<p class="govuk-body">
+  <%= govuk_button_link_to 'Invite staff user', new_staff_invitation_path %>
+</p>
